+++ extensions/vscode/webviews/homeView/src/components/views/RPackages.vue
@@ -7,14 +7,21 @@
   >
     <WelcomeView v-if="showWelcomeView">
       <template v-if="missingOrInvalidPackageFile">
+        <p>Dependencies will be detected automatically.</p>
         <p>
-          To deploy R content, you need a package file listing any package
-          dependencies, but the file ({{ rPackageFile }}) is missing, empty or
-          invalid.
-          <a class="webview-link" role="button" @click="onViewRenvDoc">
-            See the renv documentation for more details.</a
+          Alternatively, set up renv to use an isolated library and create a
+          lockfile.
+          <a class="webview-link" role="button" @click="onViewRenvDoc"
+            >See the renv documentation for more details.</a
           >
         </p>
+        <vscode-button
+          @click="onSetupRenv()"
+          :disabled="isSettingUpRenv"
+          aria-label="Set Up renv"
+        >
+          Set Up renv
+        </vscode-button>
       </template>
       <template v-if="!home.r.active.isInProject">
         <p data-automation="r-not-configured">
@@ -42,7 +49,7 @@ import TreeItem from "src/components/tree/TreeItem.vue";
 import TreeSection from "src/components/tree/TreeSection.vue";
 import WelcomeView from "src/components/WelcomeView.vue";
 
-import { computed } from "vue";
+import { computed, ref } from "vue";
 
 import { useHomeStore } from "src/stores/home";
 import { useHostConduitService } from "src/HostConduitService";
@@ -54,9 +61,18 @@ const home = useHomeStore();
 
 const hostConduit = useHostConduitService();
 
+const isSettingUpRenv = ref(false);
+
 const onRefresh = () => {
   hostConduit.sendMsg({
     kind: WebviewToHostMessageType.REFRESH_R_PACKAGES,
+  });
+};
+
+const onSetupRenv = () => {
+  isSettingUpRenv.value = true;
+  hostConduit.sendMsg({
+    kind: WebviewToHostMessageType.SCAN_R_PACKAGE_REQUIREMENTS,
   });
 };